refactor(utils): move TTL cache source files into /cache directory
The source files were living at the package root while the barrel sat
in cache/. Mirrors the structure of every other sub-entry and removes
the surprise of finding ../ttl-cache.js imports referenced from
within cache/index.ts.

No behaviour change.
