+++ packages/untp-utils/src/cache/index.ts
@@ -1,2 +1,2 @@
-export type { TtlCache, TtlCacheOptions } from '../ttl-cache.js';
-export { makeInMemoryTtlCache } from '../in-memory-ttl-cache.js';
+export type { TtlCache, TtlCacheOptions } from './ttl-cache.js';
+export { makeInMemoryTtlCache } from './in-memory-ttl-cache.js';
